Move all conan requirements to separate `conanfile.txt`s
This change removes the CONAN_EXTRA_REQUIRES variable, and moves all
conditional Conan requirements to separate `conanfile.txt` files, local
to the directories where the dependency is needed.
These requirements are installed when `add_subdirectory()` is called
on a directory that includes a `conanfile.txt` file.
This way, `sdl2` is only installed when CMake enters the `src/sdl` directory.

This helps to separate the dependencies of the `leaf node` example directories
from the `root node` top level directory. The leaves should know all the
details about themselves that the root doesn't need to know.